naming and data type fixes

+++ Shared/ComponentInstance.cs
@@ -16,7 +16,7 @@ namespace ADArCWebApp.Shared
 
 		public Dictionary<int, List<InstanceConnection>> connMap = new(); //this.pinId -> connection
 		public node gsNode;
-		public int midSignal;
+		public long midSignal;
 		public Dictionary<string, object> compParams = new Dictionary<string, object>();
 
 		public ComponentInstance(int globalId, node gsNode, double x = 10.0, double y = 10.0) {
